Remove storage state
