+++ src/eth/storage/inmemory/inmemory_temporary.rs
@@ -7,13 +7,14 @@ use tokio::sync::RwLock;
 use tokio::sync::RwLockReadGuard;
 use tokio::sync::RwLockWriteGuard;
 
+use super::inmemory_account::InMemoryAccount;
 use super::inmemory_account::InMemoryAccountTemporary;
-use super::inmemory_state::StorageState;
 use crate::eth::primitives::Account;
 use crate::eth::primitives::Address;
 use crate::eth::primitives::BlockNumber;
 use crate::eth::primitives::Execution;
 use crate::eth::primitives::ExecutionConflicts;
+use crate::eth::primitives::ExecutionConflictsBuilder;
 use crate::eth::primitives::Slot;
 use crate::eth::primitives::SlotIndex;
 use crate::eth::primitives::StoragePointInTime;
@@ -27,16 +28,6 @@ pub struct InMemoryStorageTemporary {
 #[derive(Debug, Default)]
 struct InMemoryTemporaryStorageState {
     accounts: HashMap<Address, InMemoryAccountTemporary>,
-}
-
-impl StorageState<InMemoryAccountTemporary> for InMemoryTemporaryStorageState {
-    fn get_accounts(&self) -> &HashMap<Address, InMemoryAccountTemporary> {
-        &self.accounts
-    }
-
-    fn get_accounts_mut(&mut self) -> &mut HashMap<Address, InMemoryAccountTemporary> {
-        &mut self.accounts
-    }
 }
 
 impl InMemoryStorageTemporary {
@@ -58,8 +49,41 @@ impl TemporaryStorage for InMemoryStorageTemporary {
     // ------------------------------------------------------------------------
 
     async fn check_conflicts(&self, execution: &Execution) -> anyhow::Result<Option<ExecutionConflicts>> {
-        let state_lock = self.lock_read().await;
-        Ok(state_lock.check_conflicts(execution))
+        let state = self.lock_read().await;
+        let mut conflicts = ExecutionConflictsBuilder::default();
+
+        for change in &execution.changes {
+            let address = &change.address;
+
+            if let Some(account) = state.accounts.get(address) {
+                // check account info conflicts
+                if let Some(touched_nonce) = change.nonce.take_original_ref() {
+                    let nonce = account.get_current_nonce();
+                    if touched_nonce != nonce {
+                        conflicts.add_nonce(address.clone(), nonce.clone(), touched_nonce.clone());
+                    }
+                }
+                if let Some(touched_balance) = change.balance.take_original_ref() {
+                    let balance = account.get_current_balance();
+                    if touched_balance != balance {
+                        conflicts.add_balance(address.clone(), balance.clone(), touched_balance.clone());
+                    }
+                }
+
+                // check slots conflicts
+                for (touched_slot_index, touched_slot) in &change.slots {
+                    if let Some(slot) = account.get_current_slot(touched_slot_index) {
+                        if let Some(touched_slot) = touched_slot.take_original_ref() {
+                            let slot_value = slot.value.clone();
+                            if touched_slot.value != slot_value {
+                                conflicts.add_slot(address.clone(), touched_slot_index.clone(), slot_value, touched_slot.value.clone());
+                            }
+                        }
+                    }
+                }
+            }
+        }
+        Ok(conflicts.build())
     }
 
     async fn maybe_read_account(&self, address: &Address, _point_in_time: &StoragePointInTime) -> anyhow::Result<Option<Account>> {
@@ -109,8 +133,34 @@ impl TemporaryStorage for InMemoryStorageTemporary {
     }
 
     async fn save_account_changes(&self, number: BlockNumber, execution: Execution) -> anyhow::Result<()> {
-        let mut state_lock = self.lock_write().await;
-        state_lock.save_account_changes(number, execution);
+        let mut state = self.lock_write().await;
+        let is_success = execution.is_success();
+        for changes in execution.changes {
+            let account = state
+                .accounts
+                .entry(changes.address.clone())
+                .or_insert_with(|| InMemoryAccountTemporary::new(changes.address));
+
+            // account basic info
+            if let Some(nonce) = changes.nonce.take() {
+                account.set_nonce(number, nonce);
+            }
+            if let Some(balance) = changes.balance.take() {
+                account.set_balance(number, balance);
+            }
+            if let Some(Some(bytecode)) = changes.bytecode.take() {
+                account.set_bytecode(number, bytecode);
+            }
+
+            // slots
+            if is_success {
+                for (_, slot) in changes.slots {
+                    if let Some(slot) = slot.take_modified() {
+                        account.set_slot(number, slot);
+                    }
+                }
+            }
+        }
         Ok(())
     }
